feat(config-helper): extracts loader into separate files
