+++ packages/config-helper/src/lib/loader/default.loader.ts
@@ -1,0 +1,7 @@
+import { NormalizedConfigDefinition } from '../normalized-schema'
+import { Loader } from './loader'
+import { ok } from 'neverthrow'
+
+export const defaultLoader: Loader = <T>(environment: NodeJS.ProcessEnv, configDef: NormalizedConfigDefinition<T>) => {
+	return ok(configDef.defaultValue)
+}
